Fixes
Teleporters are now working properly.
Loading screens are now included for Scene->Scene teleportation.
CustomObjectPool is included. It's a copy of the DefaultObjectPool with minor adjustments. Objects are no longer activated instantly and we use IPooledResettable to reset components that contain the interface.
Database.cfg should now be created/loaded from the proper location.