Fixed object pooling related bugs
Object pooling was not properly pooling objects on the scene server. This was because there sometimes was an ID conflict between the connection on the world server, and a game client connected to the scene server.

Pooling also will not generate an exception when a observerCondition tries to do a check on a reused pool object. The condition did not have a correctly working NetworkObject reference. This is fixed in FishNet 3.7 due to some NetworkObserver refactoring.