Bug Fix in OnDestroy.
Fixed a bug involving the Client Singleton. The singleton has been removed completely. You will now need to inject the dependency when you require access to the Client. A commented area in the Client.cs Awake function denotes where dependency injection should take place.

The application will now exit if the NetworkManager or LoginAuthenticator cannot be found in the Client.