AbilityController v0.1
Pre-test version of the new CSP AbilityController.
This commit contains the primary functionality of the new Ability system.
More events are on the way.
Synchronization of configured Abilities with observers still needs to be completed.
Itemization of AbilityEvents need to be completed.

ItemContainer now contains a TryGetItem function that will validate the item slot exists and return true if it finds an item in the slot.
UIReferenceButton - UIReferenceButton.ReferenceID is now an integer.
UIDragObject - UIDragObject.ReferenceID is now an integer.