Update CustomBuildTool
CustomBuildTool now allows you to Build the Client and Server in new ways.
You can now build the Windows OR Linux client and servers with one button.
Separated the Server Setup tools to a new directory for ease of use.
ServerLauncher will now attempt to load the first Server scene available if no parameters are given. This allows you to ignore the launcher scripts for singular server builds.