Fix Naming Bug, Added UITarget, Added MaxCharacters, Added UIAbilities
Fixed a bug with event assignment in Dictionary. Naming system is working properly now.
Added a Target window that will display your current target. It uses the TargetController LayerMask to target objects.
Added a Max Character count to CharacterCreateSystem. Default is 8.
Added the Abilities UI. This is currently in progress.
Fixed a few broadcast naming inconsistencies.
Adjusted the default UI. It is now a darker theme and easier on the eyes. Chat font is now default 10 instead of 9.