Add static Constants, Item Seed to DB, Directory Cleanup
Constants class contains global values used throughout the project.
Restructured the Network\Character directory.
Restructured the WorldSceneDetails directory.

Added WorldSceneSettings class which stores scene specific settings. Place the component on a GO in your World Scene to load it. Currently the only setting value is MaxClients (per scene). These values are baked with the WorldSceneDetailsCache.
Added Seed storage to items in the database.
Added OnDisconnect event in CharacterSystem. This event is invoked immediately after a character is removed from the server caches but before pooling.

Fixed an issue with the EquipmentViewCamera in the Equipment window rendering objects that were not the local character.