Fix ID sync, Add DB Update to Installer
The SetInitialValues function is now public in SyncVar.cs. This makes it easier for us to set the initial values when loading from the database.
Character.ID is now marked dirty to force synchronization on character load. This fixes the player naming issue where your name remains "Loading..."
Added a database update button to the new installer.