Add ClientLauncher and Patch Systems
Add Version, ClientExecutable, and UpdaterExecutable to Constants.
Add Patcher\PatcherGenerator which is used to generate patch meta data between two complete directories.
Add Patcher\Updater which is used to apply patch meta data to the current client version.
Add Patcher\patchserver.py which is a python web service used to reply to GET requests from the client. It also returns patch files if necessary when requested to the client.
Add ClientLauncher scene and class. The ClientLauncher class will attempt to connect to the designated web Host to retrieve the latest version. If there a version mismatch it will attempt to download a patch to bring the current version up-to-date with the latest version through the Updater.
Add Separation of WebGL and Standalone Client bootstrap scenes. The launcher is not included in WebGL builds.
Add TMPro_TextLinkHandler, this allows you to click links in your TMP_Text text fields when using <link></link>
Add HtmlAgilityPack to FishMMO-Dependencies.