Update UIControl and Authentication
Removed SrpSuccess broadcast. The Client no longer replies with an SrpSuccess broadcast when the Client session is verified.
Add SrpSuccessBroadcast. The Server will now reply to the Client with an SrpSuccessBroadcast containing the Server proof and the final authentication result.
UIControl's will now register to the UIManager immediately in Awake.
UIControl now supports OnClientSet and OnClientUnset functionality for use with Client dependency injection to the UIManager.
UIControl.OnStarting and UIControl.OnDestroying are now virtual instead of abstract.