Add Test Dungeon Instance
Adjust client to automatically unload all world scenes when a new one is set to be loaded. This ensures the character is only in one world scene at a time.
Added InstanceRotation to the database.
Fixed Character Instance ID, Position, and Rotation saving and loading.
Added wrapper for Enabling and Disabling CharacterFlags to ICharacter.
Added caching of InstanceSceneName and InstanceSceneHandle to IPlayerCharacter.