Add Failed SceneStatus
Added Failed to the SceneStatus enum. This value is assigned to the database Scene Load request entry if a Scene Server fails to load the requested scene for any reason.
World Servers will no longer attempt to load a scene that has failed loading. More fallbacks will be added soon to accommodate this functionality.
UICharacterCreate will now use race asset names instead of the prefab name for the character entity.