Add FishMMO-AppHealthMonitor
FishMMO-AppHealthMonitor is the new server launcher and health monitoring system. It is designed to monitor the status of the server ports by attempting connections to them. If the connection fails or times out it will kill the server process and restart it. AppHealthMonitor is taking over as our new dedicated server launching system.

Configure the appsettings.json to specify the executable path, process name, port, and any other details.